network_models/Unet_Depth_Esimation/requirements.txt
PyYAML==6.0.3
kiwisolver==1.4.9
urllib3==2.5.0
nvidia-cublas-cu12==12.8.4.1
contourpy==1.3.2
setuptools==59.6.0
annotated-types==0.7.0
imageio==2.37.0
GitPython==3.1.45
triton==3.4.0
numpy==2.2.6
nvidia-curand-cu12==10.3.9.90
charset-normalizer==3.4.3
nvidia-cusolver-cu12==11.7.3.90
MarkupSafe==3.0.3
tqdm==4.67.1
nvidia-nccl-cu12==2.27.3
torch==2.8.0
idna==3.10
sympy==1.14.0
pip==22.0.2
python-dateutil==2.9.0.post0
nvidia-nvjitlink-cu12==12.8.93
certifi==2025.8.3
h5py==3.14.0
smmap==5.0.2
networkx==3.4.2
matplotlib==3.10.6
nvidia-cudnn-cu12==9.10.2.21
Jinja2==3.1.6
fonttools==4.60.1
gitdb==4.0.12
torchvision==0.23.0
pillow==11.3.0
six==1.17.0
wandb==0.22.1
nvidia-cufile-cu12==1.13.1.3
path==17.1.1
typing_extensions==4.15.0
typing-inspection==0.4.2
fsspec==2025.9.0
cycler==0.12.1
pydantic==2.11.9
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cufft-cu12==11.3.3.83
sentry-sdk==2.39.0
nvidia-cuda-nvrtc-cu12==12.8.93
filelock==3.19.1
nvidia-cusparselt-cu12==0.7.1
mpmath==1.3.0
nvidia-cuda-runtime-cu12==12.8.90
protobuf==6.32.1
nvidia-nvtx-cu12==12.8.90
pyparsing==3.2.5
requests==2.32.5
platformdirs==4.4.0
pydantic_core==2.33.2
packaging==25.0
click==8.3.0
